finished mass migration from int to uint32_t.  tested with 4,000,000,000 objects.  test app used about 160 gigabytes of memory including memory allocated on the client side, the memory consumed by the red black tree, and the memory used by the memory management subsystems.  the client and red black tree both held their own copies of the client data in memory.  though in this test the client data was only 4 bytes vs the red black tree node which is 28 bytes without padding or 32 bytes when padded for 8 bytes alignment

+++ bintree/binremove_nocomments.cpp
@@ -282,7 +282,7 @@ BinaryTreeNode* BinDelete(BinaryTree* tree, BinaryTreeNode* z)
   if(y != z)
   {
     // copy all fields besides left right parent color
-    memcpy(GETCLIENT(z), GETCLIENT(y), (size_t)tree->sizeOfClient); // todo - 2nd copy in delete
+    memcpy(GETCLIENT(z), GETCLIENT(y), tree->sizeOfClient); // todo - 2nd copy in delete
   }
 
   if(Color == BLACK)
@@ -307,7 +307,7 @@ BinaryTreeNode* BinDelete(BinaryTree* tree, BinaryTreeNode* z)
 int bintree::remove(void* objectKey, binaryTreeEquivalence equalTo, void* objectResult)
 {
   BinaryTree* tree = (BinaryTree*)this;
-  
+
   int result = RETURN_ERROR;
 
   BinaryTreeNode* t = 0;
@@ -337,7 +337,7 @@ int bintree::remove(void* objectKey, binaryTreeEquivalence equalTo, void* object
   }
 
   if(objectResult)
-    memcpy(objectResult, GETCLIENT(z), (size_t)tree->sizeOfClient);
+    memcpy(objectResult, GETCLIENT(z), tree->sizeOfClient);
 
   t = BinDelete(tree, z);
 
@@ -357,7 +357,7 @@ int bintree::remove(void* objectKey, binaryTreeEquivalence equalTo, void* object
   }
 
   tree->numberOfNodes--;
-  
+
   result = RETURN_OK;
 
 error:
